% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_copy_n.R
\name{file_copy_n}
\alias{file_copy_n}
\title{Function to copy files with \strong{fs} but file by file to allow for a 
progress bar.}
\usage{
file_copy_n(path, path_new, overwrite = FALSE, progress = FALSE)
}
\arguments{
\item{path}{A character vector of paths to copy that currently exist.}

\item{path_new}{A character vector of paths to copy to.}

\item{overwrite}{If a \code{path_new} exists, should is be overwritten?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Invisible \code{path_new}
}
\description{
Function to copy files with \strong{fs} but file by file to allow for a 
progress bar.
}
