% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_microbenchmark_return.R
\name{format_microbenchmark_return}
\alias{format_microbenchmark_return}
\title{Function to format a \strong{microbenchmark} return into a tidy tibble.}
\usage{
format_microbenchmark_return(x)
}
\arguments{
\item{x}{A \strong{microbenchmark} object.}
}
\value{
Tibble.
}
\description{
Function to format a \strong{microbenchmark} return into a tidy tibble.
}
\author{
Stuart K. Grange
}
