% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cpu_package_temperature.R
\name{get_cpu_package_temperature}
\alias{get_cpu_package_temperature}
\title{Function to get CPU package temperature.}
\usage{
get_cpu_package_temperature()
}
\value{
Numeric vector with length of 1.
}
\description{
The package temperature is the most representative value of CPU temperature
as a whole.
}
\author{
Stuart K. Grange
}
