% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_device_read_speeds.R
\name{get_device_read_speeds}
\alias{get_device_read_speeds}
\title{Function to run the \code{hdparm} programme to get devices' read speeds, 
typically for hard discs.}
\usage{
get_device_read_speeds(device = "/dev/sda1", verbose = FALSE)
}
\arguments{
\item{device}{A vector of devices to get read speeds for.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
\code{get_device_read_speeds} will usually require elevated privileges to 
run.
}
\author{
Stuart K. Grange
}
