% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_find_mount_data.R
\name{get_find_mount_data}
\alias{get_find_mount_data}
\title{Function to get a system's mount point and device information.}
\usage{
get_find_mount_data()
}
\value{
Tibble.
}
\description{
Function to get a system's mount point and device information.
}
\examples{

\dontrun{
# Get mount information
get_find_mount_data()
}


}
\author{
Stuart K. Grange
}
