% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os_release_string.R
\name{get_os_release_string}
\alias{get_os_release_string}
\title{Function to return a somewhat detailed operating system release string.}
\usage{
get_os_release_string(add_prefix = TRUE)
}
\arguments{
\item{add_prefix}{Should a generic prefix (from \code{Sys.info}) be added to 
the return?}
}
\value{
Character vector with length of \code{1}.
}
\description{
Function to return a somewhat detailed operating system release string.
}
\author{
Stuart K. Grange
}
