% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_raw_to_jpg.R
\name{image_raw_to_jpg}
\alias{image_raw_to_jpg}
\title{Function to convert raw image files (usually \code{cr2}) to \code{jpg} image
files.}
\usage{
image_raw_to_jpg(
  file,
  file_output = NA,
  format = "jpg",
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{file}{Vector of raw image file names to be converted.}

\item{file_output}{If desired, a vector of output file names. If not used, 
\code{file} will be adapted.}

\item{format}{Output format of the image file. Only \code{jpg} is supported 
at this time.}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Invisible list.
}
\description{
\code{image_raw_to_jpg} requires \code{darktable-cli} to be installed.
}
\seealso{
\url{https://www.darktable.org/usermanual/en/special-topics/program-invocation/darktable-cli}
}
\author{
Stuart K. Grange
}
