% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_resize.R
\name{image_resize}
\alias{image_resize}
\title{Function to resize images with \code{ImageMagick}.}
\usage{
image_resize(file, quality = 2000, file_output = NA, verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names to be resized.}

\item{quality}{Quality in pixels. Default is 1000 which is one megapixel.}

\item{file_output}{Optional vector of output file names.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible.
}
\description{
Function to resize images with \code{ImageMagick}.
}
\author{
Stuart K. Grange
}
