% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_update_creation_date.R
\name{image_update_creation_date}
\alias{image_update_creation_date}
\title{Function to update images' creation dates.}
\usage{
image_update_creation_date(file, date, verbose = FALSE, progress = FALSE)
}
\arguments{
\item{file}{Vector of image files.}

\item{date}{Vector of dates to update \code{file}.}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Invisible \code{file}.
}
\description{
Function to update images' creation dates.
}
\author{
Stuart K. Grange.
}
