% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linux_release.R
\name{linux_release}
\alias{linux_release}
\title{Function to get Linux release information.}
\usage{
linux_release()
}
\value{
Tibble.
}
\description{
Function to get Linux release information.
}
\author{
Stuart K. Grange
}
