% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_system_loads.R
\name{log_system_loads}
\alias{log_system_loads}
\title{Function to log system logs.}
\usage{
log_system_loads(directory = NA)
}
\arguments{
\item{directory}{Directory to export data files too.}
}
\value{
Invisible, data files.
}
\description{
\code{log_system_loads} will take a measurement every five seconds and then
aggregate and export the measurements every minute.
}
\author{
Stuart K. Grange
}
