% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_combine.R
\name{pdf_combine}
\alias{pdf_combine}
\title{Function to combine \code{pdf} files using \code{pdftk}.}
\usage{
pdf_combine(file = NA, file_output = "combined_file.pdf", verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names to be combined. If not used, all files in 
working directory will be combined.}

\item{file_output}{Output file name.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible.
}
\description{
Function to combine \code{pdf} files using \code{pdftk}.
}
\author{
Stuart K. Grange
}
