% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_crop.R
\name{pdf_crop}
\alias{pdf_crop}
\title{Function to crop a pdf file.}
\usage{
pdf_crop(file, file_output = NA, verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names.}

\item{file_output}{Optional vector of output file names.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible list.
}
\description{
\code{pdf_crop} requires the PDFCROP system tool to be installed.
}
\author{
Stuart K. Grange
}
