% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_to_png.R
\name{pdf_to_png}
\alias{pdf_to_png}
\title{Function to convert a pdf file to a png file, usually used for single page 
pdf images.}
\usage{
pdf_to_png(file, file_output = NA, resolution = 320, verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names.}

\item{file_output}{Optional vector of output file names.}

\item{resolution}{Resolution of png file in DPI.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible.
}
\description{
Function to convert a pdf file to a png file, usually used for single page 
pdf images.
}
\author{
Stuart K. Grange
}
