% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/png_to_jpg.R
\name{png_to_jpg}
\alias{png_to_jpg}
\alias{png_to_jpeg}
\title{Function to convert a \code{.png} image to a \code{.jpeg} image.}
\usage{
png_to_jpg(file, file_output = NA, quality = 95, verbose = FALSE)

png_to_jpeg(file, file_output = NA, quality = 95, verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names to be resized.}

\item{file_output}{Optional vector of output file names.}

\item{quality}{Quality, represents jpeg compression (100--0). No compression 
is 100.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible \code{NULL}.
}
\description{
Function to convert a \code{.png} image to a \code{.jpeg} image.
}
\author{
Stuart K. Grange
}
