% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_version.R
\name{r_version}
\alias{r_version}
\title{Function to get R's version information.}
\usage{
r_version()
}
\value{
Data frame.
}
\description{
Function to get R's version information.
}
\author{
Stuart K. Grange
}
