% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_rhythmbox_database.R
\name{read_rhythmbox_database}
\alias{read_rhythmbox_database}
\alias{read_rhythmbox_playlists}
\title{Functions to read XML files created by the Rhythmbox music library software.}
\usage{
read_rhythmbox_database(file)

read_rhythmbox_playlists(file)
}
\arguments{
\item{file}{File name of Rhythmbox's XML database or Rhythmbox's playlist 
database.}
}
\value{
Tibble, nested in the case of \code{read_rhythmbox_database} and
standard by \code{format_rhythmbox_playlist_entry}.
}
\description{
Functions to read XML files created by the Rhythmbox music library software.
}
\author{
Stuart K. Grange
}
