% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsync_directory.R
\name{rsync_directory}
\alias{rsync_directory}
\title{Function to synchronize a local and remote directory with \code{rsync}.}
\usage{
rsync_directory(directory_local, directory_remote, file_database = NA,
  database = NA)
}
\arguments{
\item{directory_local}{Local directory to synchronize to remote directory.}

\item{directory_remote}{Remote directory to synchronize from local directory.}

\item{file_database}{Optional, database file. See \code{\link{db_connect}}.}

\item{database}{Optional, database within \code{file_database}. 
See \code{\link{db_connect}}.}
}
\value{
Invisible.
}
\description{
\code{rsync_directory} will delete files in the remote directory if not 
present in the local directory.
}
\author{
Stuart K. Grange
}
