% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_test.R
\name{speed_test}
\alias{speed_test}
\title{Function to run the \code{speedtest-cli} internet bandwidth programme.}
\usage{
speed_test(verbose = TRUE)
}
\arguments{
\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to run the \code{speedtest-cli} internet bandwidth programme.
}
\author{
Stuart K. Grange
}
