% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_quietly.R
\name{stop_quietly}
\alias{stop_quietly}
\title{Function to stop an R session without an error.}
\usage{
stop_quietly()
}
\value{
Invisible, the R session will be stopped.
}
\description{
Function to stop an R session without an error.
}
\author{
Stuart K. Grange
}
