% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_to_pdf.R
\name{svg_to_pdf}
\alias{svg_to_pdf}
\title{Function to convert a  \code{svg} file to a \code{pdf} file.}
\usage{
svg_to_pdf(file, file_output = NA, verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names.}

\item{file_output}{An optional vector of output file names.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible list containing \code{file}.
}
\description{
\code{svg_to_pdf} requires inkscape to be installed.
}
\author{
Stuart K. Grange
}
