% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_block_devices.R
\name{system_block_devices}
\alias{system_block_devices}
\title{Function to return system block devices.}
\usage{
system_block_devices()
}
\value{
Named list.
}
\description{
Function to return system block devices.
}
\author{
Stuart K. Grange
}
