% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_cpu_information.R
\name{system_cpu_information}
\alias{system_cpu_information}
\title{Function to return CPU information.}
\usage{
system_cpu_information()
}
\value{
Tibble.
}
\description{
Function to return CPU information.
}
\author{
Stuart K. Grange
}
