% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_directory_size.R
\name{system_directory_size}
\alias{system_directory_size}
\title{Function to get the total size of a directory.}
\usage{
system_directory_size(directory, unit = "kb")
}
\arguments{
\item{directory}{Vector of directories.}

\item{unit}{Vector of units. Can be \code{"kb"}, \code{"mb"}, or \code{"gb"}.}
}
\value{
Tibble.
}
\description{
\code{system_directory_size} uses the \code{du} system programme.
}
\author{
Stuart K. Grange
}
