% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_disk_space.R
\name{system_disk_space}
\alias{system_disk_space}
\title{Function to return a system's file systems' free space information.}
\usage{
system_disk_space()
}
\value{
Tibble.
}
\description{
Function to return a system's file systems' free space information.
}
\author{
Stuart K. Grange
}
