% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_essid.R
\name{system_essid}
\alias{system_essid}
\title{Function to return a system's extended service set identification (ESSID) 
names.}
\usage{
system_essid()
}
\value{
Data frame.
}
\description{
Function to return a system's extended service set identification (ESSID) 
names.
}
\author{
Stuart K. Grange
}
