% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_ip_address.R
\name{system_ip_address}
\alias{system_ip_address}
\title{Function to return system's IP address.}
\usage{
system_ip_address()
}
\value{
Character vector.
}
\description{
Function to return system's IP address.
}
\author{
Stuart K. Grange
}
