% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_last_users.R
\name{system_last_users}
\alias{system_last_users}
\title{Function to return a formatted version of the Unix \code{last} command.}
\usage{
system_last_users()
}
\value{
Tibble.
}
\description{
Function to return a formatted version of the Unix \code{last} command.
}
\author{
Stuart K. Grange
}
