% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_mac_addresses.R
\name{system_mac_addresses}
\alias{system_mac_addresses}
\title{Function to return a system's media access control (MAC) addresses.}
\usage{
system_mac_addresses(format = FALSE)
}
\arguments{
\item{format}{Should the MAC addresses be formated to be upper case and 
contain hyphens as the separator?}
}
\value{
Tibble.
}
\description{
Function to return a system's media access control (MAC) addresses.
}
\author{
Stuart K. Grange
}
