% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_memory_information.R
\name{system_memory_information}
\alias{system_memory_information}
\title{Function to return a system's memory information.}
\usage{
system_memory_information()
}
\value{
Tibble.
}
\description{
Function to return a system's memory information.
}
\author{
Stuart K. Grange
}
