% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_memory_information.R
\name{system_memory_usage}
\alias{system_memory_usage}
\title{Function to calculate percentage memory usage.}
\usage{
system_memory_usage()
}
\value{
Numeric vector with length of one.
}
\description{
Function to calculate percentage memory usage.
}
\author{
Stuart K. Grange.
}
