% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_network_connections.R
\name{system_network_connections}
\alias{system_network_connections}
\title{Function to list a system's network connections.}
\usage{
system_network_connections()
}
\value{
Tibble.
}
\description{
Function to list a system's network connections.
}
\author{
Stuart K. Grange
}
