% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_notify.R
\name{system_notify}
\alias{system_notify}
\title{Function to send a system notification to the Ubuntu desktop.}
\usage{
system_notify(urgency = "critical", title = "R: ", message = "")
}
\arguments{
\item{urgency}{Urgency of message, one of \code{"low"}, \code{"normal"}, or
\code{"critical"}. Default is \code{"critical"} because it is the most 
supported notification type.}

\item{title}{Title for notification.}

\item{message}{Message for notification.}
}
\value{
Invisible, a desktop notification.
}
\description{
Function to send a system notification to the Ubuntu desktop.
}
\author{
Stuart K. Grange
}
