% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_processes.R
\name{system_processes}
\alias{system_processes}
\title{Function to list a system's processes.}
\usage{
system_processes(by = "cpu")
}
\value{
Data frame.
}
\description{
Function to list a system's processes.
}
\author{
Stuart K. Grange
}
