% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_users_running_processes.R
\name{system_running_processes}
\alias{system_running_processes}
\title{Function to count number of processes running.}
\usage{
system_running_processes()
}
\value{
Integer vector with length of 1.
}
\description{
Function to count number of processes running.
}
\author{
Stuart K. Grange
}
