% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_sensors.R
\name{system_sensors}
\alias{system_sensors}
\title{Function to format \code{sensors} output.}
\usage{
system_sensors()
}
\value{
Data frame.
}
\description{
Function to format \code{sensors} output.
}
\author{
Stuart K. Grange
}
