% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_time_zone.R
\name{system_time_zone}
\alias{system_time_zone}
\title{Function to get system's time zone string.}
\usage{
system_time_zone()
}
\value{
Character vector with length of one.
}
\description{
Function to get system's time zone string.
}
\author{
Stuart K. Grange
}
