% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_usb_devices.R
\name{system_usb_devices}
\alias{system_usb_devices}
\title{Function to return a system's USB devices from a \code{lsusb} system call.}
\usage{
system_usb_devices()
}
\value{
Tibble.
}
\description{
Function to return a system's USB devices from a \code{lsusb} system call.
}
\author{
Stuart K. Grange
}
