% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_users_running_processes.R
\name{system_users_running_processes}
\alias{system_users_running_processes}
\title{Function to list or count users who have running processes.}
\usage{
system_users_running_processes(count = FALSE)
}
\value{
Character or integer vector.
}
\description{
Function to list or count users who have running processes.
}
\author{
Stuart K. Grange
}
