% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubuntu_release.R
\name{ubuntu_release}
\alias{ubuntu_release}
\title{Function to get Ubuntu release information.}
\usage{
ubuntu_release()
}
\value{
Data frame.
}
\description{
Function to get Ubuntu release information.
}
\author{
Stuart K. Grange
}
