% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ubuntu_test.R
\name{ubuntu_test}
\alias{ubuntu_test}
\title{Function to test if the current system is an Ubuntu distribution.}
\usage{
ubuntu_test()
}
\value{
Logical vector with length of 1.
}
\description{
Function to test if the current system is an Ubuntu distribution.
}
