% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uptime.R
\name{uptime}
\alias{uptime}
\title{Function to get and format the \code{uptime} system call.}
\usage{
uptime(round = FALSE)
}
\arguments{
\item{round}{Should the dates and times be rounded to the nearest second?}
}
\description{
Function to get and format the \code{uptime} system call.
}
\author{
Stuart K. Grange
}
