% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webp_to_jpg.R
\name{webp_to_jpg}
\alias{webp_to_jpg}
\title{Function to convert a \code{.webp} image to a \code{.jpg} image.}
\usage{
webp_to_jpg(file, file_output = NA, verbose = FALSE)
}
\arguments{
\item{file}{Vector of file names to be converted.}

\item{file_output}{Optional vector of output file names.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible \code{file}.
}
\description{
Function to convert a \code{.webp} image to a \code{.jpg} image.
}
\author{
Stuart K. Grange
}
