% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_resize.R
\name{image_resize}
\alias{image_resize}
\title{Function to resize images with \code{ImageMagick}.}
\usage{
image_resize(
  file,
  file_output = NA,
  quality = 2000,
  verbose = FALSE,
  progress = FALSE
)
}
\arguments{
\item{file}{Vector of file names to be resized.}

\item{file_output}{Optional vector of output file names.}

\item{quality}{Quality in pixels. Default is 2000 which is two megapixels.}

\item{verbose}{Should the function give messages?}

\item{progress}{Should a progress bar be displayed?}
}
\value{
Invisible list.
}
\description{
Function to resize images with \code{ImageMagick}.
}
\author{
Stuart K. Grange
}
