% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_update_creation_date.R
\name{image_update_creation_date}
\alias{image_update_creation_date}
\title{Function to update an image's creation dates.}
\usage{
image_update_creation_date(file, date, verbose = FALSE)
}
\arguments{
\item{file}{File name.}

\item{date}{Creation date to update \code{file} with.}

\item{verbose}{Should the function give messages?}
}
\value{
Invisible \code{file}.
}
\description{
Function to update an image's creation dates.
}
\author{
Stuart K. Grange.
}
