% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r_version.R
\name{r_version}
\alias{r_version}
\title{Function to return a formatred \code{R.Version} call.}
\usage{
r_version()
}
\value{
Tibble with one row.
}
\description{
Function to return a formatred \code{R.Version} call.
}
\author{
Stuart K. Grange
}
