% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed_test.R
\name{speed_test}
\alias{speed_test}
\title{Function to run the \code{speedtest-cli} internet bandwidth programme.}
\usage{
speed_test(n = 1, sleep = NA, verbose = TRUE)
}
\arguments{
\item{n}{Number of times to run the speed test.}

\item{sleep}{If \code{n} is not 1, the number of seconds to sleep before 
speed tests.}

\item{verbose}{Should the function give messages?}
}
\value{
Tibble.
}
\description{
Function to run the \code{speedtest-cli} internet bandwidth programme.
}
\author{
Stuart K. Grange
}
