% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synchronize_directory.R
\name{synchronize_directory}
\alias{synchronize_directory}
\title{Function to synchronize a local and remote directory with \code{rsync}.}
\usage{
synchronize_directory(directory_local, directory_remote, verbose = TRUE)
}
\arguments{
\item{directory_local}{Local directory to synchronize to remote directory.}

\item{directory_remote}{Remote directory to synchronize from local directory.}

\item{verbose}{Should the function give messages?}
}
\value{
Data frame with programme timing information.
}
\description{
\code{synchronize_directory} will delete files in the remote directory if not 
present in the local directory.
}
\author{
Stuart K. Grange
}
