% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_info.R
\name{system_info}
\alias{system_info}
\title{Function to return a better formatted \code{Sys.info} call.}
\usage{
system_info()
}
\value{
Tibble.
}
\description{
Function to return a better formatted \code{Sys.info} call.
}
\author{
Stuart K. Grange
}
