% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_notify.R
\name{system_notify}
\alias{system_notify}
\title{Function to send a system notification to the Ubuntu desktop.}
\usage{
system_notify(title = "R", message = "")
}
\arguments{
\item{title}{Title for notification.}

\item{message}{Message for notification.}
}
\value{
Invisible.
}
\description{
Function to send a system notification to the Ubuntu desktop.
}
\author{
Stuart K. Grange
}
