% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_uptime.R
\name{system_uptime}
\alias{system_uptime}
\title{Function to get system uptime.}
\usage{
system_uptime(round = TRUE)
}
\arguments{
\item{round}{Should the dates and times be rounded to the nearest second?}
}
\description{
Function to get system uptime.
}
\author{
Stuart K. Grange
}
